+++ include/zephyr/net/websocket.h
@@ -16,8 +16,8 @@
 #include <zephyr/kernel.h>
 
 #include <zephyr/net/net_ip.h>
-#include <zephyr/net/http_parser.h>
-#include <zephyr/net/http_client.h>
+#include <zephyr/net/http/parser.h>
+#include <zephyr/net/http/client.h>
 
 #ifdef __cplusplus
 extern "C" {
