+++ samples/net/cloud/tagoio_http_post/src/main.c
@@ -9,7 +9,7 @@ LOG_MODULE_REGISTER(tagoio_http_post, CONFIG_TAGOIO_HTTP_POST_LOG_LEVEL);
 
 #include <zephyr/kernel.h>
 #include <zephyr/net/socket.h>
-#include <zephyr/net/http_client.h>
+#include <zephyr/net/http/client.h>
 #include <zephyr/random/rand32.h>
 #include <stdio.h>
