+++ samples/net/cloud/tagoio_http_post/src/sockets.c
@@ -12,7 +12,7 @@ LOG_MODULE_DECLARE(tagoio_http_post, CONFIG_TAGOIO_HTTP_POST_LOG_LEVEL);
 #include <zephyr/net/socketutils.h>
 #include <zephyr/net/dns_resolve.h>
 #include <zephyr/net/tls_credentials.h>
-#include <zephyr/net/http_client.h>
+#include <zephyr/net/http/client.h>
 
 #include "sockets.h"
